test: split location filter strict coverage
(cherry picked from commit 07e1aa702e795f39f15f6c4c4e3daccde239ecbd)
